+++ tests/ChunkTest.php
@@ -22,42 +22,42 @@ final class ChunkTest extends TestCase
      */
     private $chunk;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->chunk = new Chunk;
     }
 
-    public function testCanBeCreatedWithoutArguments()
+    public function testCanBeCreatedWithoutArguments(): void
     {
         $this->assertInstanceOf(Chunk::class, $this->chunk);
     }
 
-    public function testStartCanBeRetrieved()
+    public function testStartCanBeRetrieved(): void
     {
         $this->assertSame(0, $this->chunk->getStart());
     }
 
-    public function testStartRangeCanBeRetrieved()
+    public function testStartRangeCanBeRetrieved(): void
     {
         $this->assertSame(1, $this->chunk->getStartRange());
     }
 
-    public function testEndCanBeRetrieved()
+    public function testEndCanBeRetrieved(): void
     {
         $this->assertSame(0, $this->chunk->getEnd());
     }
 
-    public function testEndRangeCanBeRetrieved()
+    public function testEndRangeCanBeRetrieved(): void
     {
         $this->assertSame(1, $this->chunk->getEndRange());
     }
 
-    public function testLinesCanBeRetrieved()
+    public function testLinesCanBeRetrieved(): void
     {
         $this->assertSame([], $this->chunk->getLines());
     }
 
-    public function testLinesCanBeSet()
+    public function testLinesCanBeSet(): void
     {
         $this->assertSame([], $this->chunk->getLines());