+++ tests/LineTest.php
@@ -22,22 +22,22 @@ final class LineTest extends TestCase
      */
     private $line;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->line = new Line;
     }
 
-    public function testCanBeCreatedWithoutArguments()
+    public function testCanBeCreatedWithoutArguments(): void
     {
         $this->assertInstanceOf(Line::class, $this->line);
     }
 
-    public function testTypeCanBeRetrieved()
+    public function testTypeCanBeRetrieved(): void
     {
         $this->assertSame(Line::UNCHANGED, $this->line->getType());
     }
 
-    public function testContentCanBeRetrieved()
+    public function testContentCanBeRetrieved(): void
     {
         $this->assertSame('', $this->line->getContent());
     }