+++ tests/Output/AbstractChunkOutputBuilderTest.php
@@ -30,7 +30,7 @@ final class AbstractChunkOutputBuilderTest extends TestCase
      *
      * @dataProvider provideGetCommonChunks
      */
-    public function testGetCommonChunks(array $expected, string $from, string $to, int $lineThreshold = 5)
+    public function testGetCommonChunks(array $expected, string $from, string $to, int $lineThreshold = 5): void
     {
         $output = new class extends AbstractChunkOutputBuilder {
             public function getDiff(array $diff): string