+++ tests/Utils/UnifiedDiffAssertTraitTest.php
@@ -24,7 +24,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
      *
      * @dataProvider provideValidCases
      */
-    public function testValidCases(string $diff)
+    public function testValidCases(string $diff): void
     {
         $this->assertValidUnifiedDiffFormat($diff);
     }
@@ -57,7 +57,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
         ];
     }
 
-    public function testNoLinebreakEnd()
+    public function testNoLinebreakEnd(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Expected diff to end with a line break, got "C".', '#')));
@@ -65,7 +65,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
         $this->assertValidUnifiedDiffFormat("A\nB\nC");
     }
 
-    public function testInvalidStartWithoutHeader()
+    public function testInvalidStartWithoutHeader(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Expected line to start with '@', '-' or '+', got \"A\n\". Line 1.", '#')));
@@ -73,7 +73,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
         $this->assertValidUnifiedDiffFormat("A\n");
     }
 
-    public function testInvalidStartHeader1()
+    public function testInvalidStartHeader1(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Line 1 indicates a header, so line 2 must start with \"+++\".\nLine 1: \"--- A\n\"\nLine 2: \"+ 1\n\".", '#')));
@@ -81,7 +81,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
         $this->assertValidUnifiedDiffFormat("--- A\n+ 1\n");
     }
 
-    public function testInvalidStartHeader2()
+    public function testInvalidStartHeader2(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Header line does not match expected pattern, got \"+++ file	X\n\". Line 2.", '#')));
@@ -89,7 +89,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
         $this->assertValidUnifiedDiffFormat("--- A\n+++ file\tX\n");
     }
 
-    public function testInvalidStartHeader3()
+    public function testInvalidStartHeader3(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Date of header line does not match expected pattern, got "[invalid date]". Line 1.', '#')));
@@ -104,7 +104,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testInvalidStartHeader4()
+    public function testInvalidStartHeader4(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Expected header line to start with \"+++  \", got \"+++INVALID\n\". Line 2.", '#')));
@@ -119,7 +119,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testInvalidLine1()
+    public function testInvalidLine1(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Expected line to start with '@', '-' or '+', got \"1\n\". Line 5.", '#')));
@@ -134,7 +134,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testInvalidLine2()
+    public function testInvalidLine2(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Expected string length of minimal 2, got 1. Line 4.', '#')));
@@ -148,7 +148,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testHunkInvalidFormat()
+    public function testHunkInvalidFormat(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote("Hunk header line does not match expected pattern, got \"@@ INVALID -1,1 +1,1 @@\n\". Line 3.", '#')));
@@ -162,7 +162,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testHunkOverlapFrom()
+    public function testHunkOverlapFrom(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected new hunk; "from" (\'-\') start overlaps previous hunk. Line 6.', '#')));
@@ -179,7 +179,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testHunkOverlapTo()
+    public function testHunkOverlapTo(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected new hunk; "to" (\'+\') start overlaps previous hunk. Line 6.', '#')));
@@ -196,7 +196,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testExpectHunk1()
+    public function testExpectHunk1(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Expected hunk start (\'@\'), got "+". Line 6.', '#')));
@@ -211,7 +211,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testExpectHunk2()
+    public function testExpectHunk2(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected hunk start (\'@\'). Line 6.', '#')));
@@ -226,7 +226,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testMisplacedLineAfterComments1()
+    public function testMisplacedLineAfterComments1(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected line as 2 "No newline" markers have found, ". Line 8.', '#')));
@@ -243,7 +243,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testMisplacedLineAfterComments2()
+    public function testMisplacedLineAfterComments2(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected line as 2 "No newline" markers have found, ". Line 7.', '#')));
@@ -259,7 +259,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testMisplacedLineAfterComments3()
+    public function testMisplacedLineAfterComments3(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected line as 2 "No newline" markers have found, ". Line 7.', '#')));
@@ -275,7 +275,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testMisplacedComment()
+    public function testMisplacedComment(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected "\ No newline at end of file", it must be preceded by \'+\' or \'-\' line. Line 1.', '#')));
@@ -285,7 +285,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testUnexpectedDuplicateNoNewLineEOF()
+    public function testUnexpectedDuplicateNoNewLineEOF(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected "\\ No newline at end of file", "\\" was already closed. Line 8.', '#')));
@@ -302,7 +302,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testFromAfterClose()
+    public function testFromAfterClose(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Not expected from (\'-\'), already closed by "\ No newline at end of file". Line 6.', '#')));
@@ -318,7 +318,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testSameAfterFromClose()
+    public function testSameAfterFromClose(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Not expected same (\' \'), \'-\' already closed by "\ No newline at end of file". Line 6.', '#')));
@@ -334,7 +334,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testToAfterClose()
+    public function testToAfterClose(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Not expected to (\'+\'), already closed by "\ No newline at end of file". Line 6.', '#')));
@@ -350,7 +350,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testSameAfterToClose()
+    public function testSameAfterToClose(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Not expected same (\' \'), \'+\' already closed by "\ No newline at end of file". Line 6.', '#')));
@@ -366,7 +366,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testUnexpectedEOFFromMissingLines()
+    public function testUnexpectedEOFFromMissingLines(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected EOF, number of lines in hunk "from" (\'-\')) mismatched. Line 7.', '#')));
@@ -381,7 +381,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testUnexpectedEOFToMissingLines()
+    public function testUnexpectedEOFToMissingLines(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected EOF, number of lines in hunk "to" (\'+\')) mismatched. Line 7.', '#')));
@@ -396,7 +396,7 @@ final class UnifiedDiffAssertTraitTest extends TestCase
 ');
     }
 
-    public function testUnexpectedEOFBothFromAndToMissingLines()
+    public function testUnexpectedEOFBothFromAndToMissingLines(): void
     {
         $this->expectException(\UnexpectedValueException::class);
         $this->expectExceptionMessageRegExp(\sprintf('#^%s$#', \preg_quote('Unexpected EOF, number of lines in hunk "from" (\'-\')) and "to" (\'+\') mismatched. Line 7.', '#')));