+++ ChangeLog.md
@@ -2,7 +2,17 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## [2.1.0] - 2017-MM-DD
+## [3.0.0] - 2018-MM-DD
+
+* The `StrictUnifiedDiffOutputBuilder` implementation of the `DiffOutputBuilderInterface` was added
+
+### Changed
+
+* The default `DiffOutputBuilderInterface` implementation now generates context lines (unchanged lines)
+
+### Removed
+
+* Removed support for PHP 7.0
 
 ### Fixed
 
@@ -24,6 +34,6 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 * This component is no longer supported on PHP 5.6
 
-[2.1.0]: https://github.com/sebastianbergmann/diff/compare/2.0...2.1.0
+[3.0.0]: https://github.com/sebastianbergmann/diff/compare/2.0...3.0.0
 [2.0.1]: https://github.com/sebastianbergmann/diff/compare/c341c98ce083db77f896a0aa64f5ee7652915970...2.0.1
 [2.0.0]: https://github.com/sebastianbergmann/diff/compare/1.4...c341c98ce083db77f896a0aa64f5ee7652915970